New test, more comments.
Made a new test to more fit the assignment.
Had a thought at 3 in the morning that I could just change the addtolist return to include the content property of the response class to get access to the list in the test to be able to make sure the product was in it.